refactor(control-plane): complete init/summary handler boundaries
